Hello, Home Phrase
Switch Activates Hello, Home Phrase

Ties a Hello, Home phrase to a switch's (virtual or real) on/off state. Perfect for use with IFTTT.
Simple define a switch to be used, then tie the on/off state of the switch to a specific Hello, Home phrases.
Connect the switch to an IFTTT action, and the Hello, Home phrase will fire with the switch state change.