Switch Changes Mode
Ties a mode to a switch's (virtual or real) on/off state. Perfect for use with IFTTT.
Simple define a switch to be used, then tie the on/off state of the switch to a specific mode.
Connect the switch to an IFTTT action, and the mode will fire with the switch state change.